+++ ChangeLog.rst
@@ -1,6 +1,48 @@
 ChangeLog
 =========
 
+Version 1.7.0_ - 2018-12-09
+---------------------------
+
+* [docs] Fix the owned/starred usage documentation
+* [docs] Add a warning about http to https redirects
+* Fix the https redirection test
+* [docs] Add a note about GroupProject limited API
+* Add missing comma in ProjectIssueManager _create_attrs
+* More flexible docker image
+* Add project protected tags management
+* [cli] Print help and usage without config file
+* Rename MASTER_ACCESS to MAINTAINER_ACCESS
+* [docs] Add docs build information
+* Use docker image with current sources
+* [docs] Add PyYAML requirement notice
+* Add Gitter badge to README
+* [docs] Add an example of pipeline schedule vars listing
+* [cli] Exit on config parse error, instead of crashing
+* Add support for resource label events
+* [docs] Fix the milestone filetring doc (iid -> iids)
+* [docs] Fix typo in custom attributes example
+* Improve error message handling in exceptions
+* Add support for members all() method
+* Add access control options to protected branch creation
+
+Version 1.6.0_ - 2018-08-25
+---------------------------
+
+* [docs] Don't use hardcoded values for ids
+* [docs] Improve the snippets examples
+* [cli] Output: handle bytes in API responses
+* [cli] Fix the case where we have nothing to print
+* Project import: fix the override_params parameter
+* Support group and global MR listing
+* Implement MR.pipelines()
+* MR: add the squash attribute for create/update
+* Added support for listing forks of a project
+* [docs] Add/update notes about read-only objects
+* Raise an exception on https redirects for PUT/POST
+* [docs] Add a FAQ
+* [cli] Fix the project-export download
+
 Version 1.5.1_ - 2018-06-23
 ---------------------------
 
@@ -643,7 +685,9 @@ Version 0.1 - 2013-07-08
 
 * Initial release
 
-.. _1.5.1: https://github.com/python-gitlab/python-gitlab/compare/1.4.0...1.5.1
+.. _1.7.0: https://github.com/python-gitlab/python-gitlab/compare/1.6.0...1.7.0
+.. _1.6.0: https://github.com/python-gitlab/python-gitlab/compare/1.5.1...1.6.0
+.. _1.5.1: https://github.com/python-gitlab/python-gitlab/compare/1.5.0...1.5.1
 .. _1.5.0: https://github.com/python-gitlab/python-gitlab/compare/1.4.0...1.5.0
 .. _1.4.0: https://github.com/python-gitlab/python-gitlab/compare/1.3.0...1.4.0
 .. _1.3.0: https://github.com/python-gitlab/python-gitlab/compare/1.2.0...1.3.0